# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API specification
  description: |
    This OpenAPI document describes the API for a messaging web application.
  version: "0.0.1"
...
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: The user name that conforms to the format
              properties:
                name:
                  type: string
                  description: The display name used for login or rigistration
                  example: Maria
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
              required: 
                - name
        required: true
      responses:
        "201":
          description: User login action successful
          content:
            application/json:
              schema:
                type: object
                description: User ID
                properties:
                  identifier:
                    type: integer
                    description: The unique ID to be used in the Authorization
                    example: 1001
        "400":
          description: The length of the name is out of range
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /user/name:
    put:
      tags: ["user"]
      summary: Update the user's name
      description: |
        If the user is authenticated and the name is not duplicated with others,
        the name will be changed
      operationId: setMyUserName
      requestBody:
        description: User name
        content:
          application/json:
            schema:
              type: object
              description: The user name that conforms to the format
              properties:
                newName:
                  type: string
                  description: The new display name for the user
                  example: Maria
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
              required:
                - newName
        required: true
      responses:
        "200":
          description: User name updated successfully
          content: {}
        "400":
          description: The length of the name is out of range
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "409":
          description: The name is already exist
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /user/photo:
    put: 
      tags: ["user"]
      summary: Set user photo
      description: If the user is authenticated, the photo will be changed
      operationId: setMyPhoto
      requestBody:
        description: User photo
        content: 
          multipart/form-data:
            schema:
              type: object
              description: User's profile photo file
              properties:
                newPhoto:
                  type: string
                  format: binary
                  description: The profile photo file to upload
              required:
                - newPhoto
        required: true
      responses:
        "200":
          description: User photo updated successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /user/me:
    get:
      tags: ["user"]
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      operationId: getMyProfile
      responses:
        "200":
          description: Successfully retrieved user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user-info"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /users:
    get:
      tags: ["user"]
      summary: List users
      description: List users with optional search query
      operationId: listUsers
      parameters:
        - name: q
          in: query
          required: false
          description: Search query for username
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user-info"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  /messages:
    post:
      tags: ["message"]
      summary: Send message
      description: |
        Send a new message(text or photo) to the conversation 
        or reply an old message
      operationId: sendMessage
      requestBody:
        description: message details
        content:
          application/json:
            schema:
              type: object
              description: Information about the message
              properties:
                conversationId:
                  type: integer
                  description: The ID of the target conversation
                content:
                  type: string
                  description: The text content or URL for the photo
                  minLength: 1
                  maxLength: 200
                contentType:
                  type: string
                  enum:
                    - text
                    - photo
                  description: Type pf content being sent
                replyToId:
                  type: integer
                  nullable: true
                  description: Optional ID of the message being replied to
              required:
                - conversationId
                - content
                - contentType
        required: true
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/message"
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Target conversation not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        required: true
        description: this is the message id
        schema:
          type: integer
          description: an incremental number 
          
    delete:
      tags: ["message"]
      summary: Delete message
      description: Only the sender can choose to delete the message he sent
      operationId: deleteMessage
      responses:
        "200":
          description: The message was successfullt deleted
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "403":
          description: No permission to delete this message
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Message not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /messages/{messageId}/forward:
    parameters:
      - name: messageId
        in: path
        required: true
        description: this is the message id
        schema:
          type: integer
          description: an incremental number 
          
    post:
      tags: ["message"]
      summary: Forward message
      description: Forward a message to one or more other conversations
      operationId: forwardMessage
      requestBody:
        description: target conversation details
        content:
          application/json:
            schema:
              type: object
              description: Array of target conversation details
              properties:
                targetConversationIds:
                  type: array
                  description: List of conversation IDs to forward the message to
                  minItems: 1
                  maxItems: 10
                  items:
                    type: integer
              required:
                - targetConversationIds
        required: true
      responses:
        "200":
          description: Message forwarding successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Original message or target conversationn is not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
  /messages/{messageId}/reaction:
    parameters:
      - name: messageId
        in: path
        required: true
        description: this is the message id
        schema:
          type: integer
          description: an incremental number 
          
    post:
      tags: ["reaction"]
      summary: Comment message
      description: Add or update a reaction to the message
      operationId: commentMessage
      requestBody:
        description: Reaction emotion
        content:
          application/json:
            schema:
              type: object
              description: The emoticon info
              properties:
                emoticon:
                  type: string
                  description: The reaction emotion
                  minLength: 1
                  maxLength: 1
              required: 
                - emoticon
        required: true
      responses:
        "200":
          description: Reaction updated successfully
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Original message or target conversationn is not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
          
    delete:
      tags: ["reaction"]
      summary: Uncomment message
      description: Cancel the comment to the message
      operationId: uncommentMessage
      responses:
        "200":
          description: The reaction was successfullt deleted
          content: {}
        "401":
          description: The user is unauthorized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
        "404":
          description: Original message or target conversationn is not found
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorResponse"}
...
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: WASAText-ID
      description: |
        Authentication is performed by passing the user's ID 
        in the Authorization header as a Bearer token
  schemas:
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        error:
          type: string
          description: A brief description of error
          minLength: 0
          maxLength: 200
        code:
          type: integer
          description: Optional specific error code
      required: 
        - error
    user-info:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: string
      required: [id, name]
...
    reaction:
      type: object
      description: A emotion added to a message
      properties:
        reactorName: 
          type: string
          description: The name of the user who added the reaction
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
        emoticon:
          type: string
          description: The emotion
          minLength: 1
          maxLength: 1
      required: 
        - reactorName
        - emoticon
    message:
      type: object
      description: A complete message object, used in responses
      properties:
        id:
          type: integer
          description: The unique ID of the message
        conversationId:
          type: integer
          description: The  ID of the conversation the message belongs to
        senderId:
          type: integer
          description: The ID of the message sender
        senderName:
          type: string
          description: The display name of message sender
          minLength: 1
          maxLength: 16
          pattern: '^.*?$'
        timeStamp:
          type: string
          format: date-time
          description: The time that message was sent
        content:
          type: string
          description: The message content details
          minLength: 1
          maxLength: 200
        contentType:
          type: string
          enum:
            - text
            - photo
          description: The type of the content
        replyToId:
          type: integer
          nullable: true
          description: The ID of the message that replying to
        status:
          type: integer
          description: "0: Sent, 1: Received by all, 2: Read by all."
          enum:
            - 0
            - 1
            - 2
        reactions:
          type: array
          description: List of reactions to this message
          minItems: 0
          maxItems: 10
          items: 
            $ref: '#/components/schemas/reaction'
      required:
        - id
        - conversationId
        - senderId
        - senderName
        - timeStamp
        - content
        - contentType
        - status
...
security:
  - bearerAuth: []